+++ src/main/java/org/halvors/nuclearphysics/client/gui/component/GuiTemperatureInfo.java
@@ -4,8 +4,8 @@ import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 import org.halvors.nuclearphysics.client.gui.IGuiWrapper;
 import org.halvors.nuclearphysics.common.ConfigurationManager.General;
+import org.halvors.nuclearphysics.common.science.unit.EnumTemperatureUnit;
 import org.halvors.nuclearphysics.common.type.EnumResource;
-import org.halvors.nuclearphysics.common.unit.EnumTemperatureUnit;
 import org.halvors.nuclearphysics.common.utility.LanguageUtility;
 import org.halvors.nuclearphysics.common.utility.ResourceUtility;
