+++ src/main/java/org/halvors/nuclearphysics/client/render/block/reactor/RenderThermometer.java
@@ -5,10 +5,10 @@ import net.minecraft.client.renderer.tileentity.TileEntitySpecialRenderer;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 import org.halvors.nuclearphysics.client.utility.RenderUtility;
+import org.halvors.nuclearphysics.common.science.unit.UnitDisplay;
 import org.halvors.nuclearphysics.common.tile.reactor.TileThermometer;
 import org.halvors.nuclearphysics.common.type.EnumColor;
 import org.halvors.nuclearphysics.common.type.Position;
-import org.halvors.nuclearphysics.common.unit.UnitDisplay;
 
 @SideOnly(Side.CLIENT)
 public class RenderThermometer extends TileEntitySpecialRenderer<TileThermometer> {
